+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
   * Add authz grant/revoke command to `marker` cli
 * Add authz and feegrant modules [PR 384](https://github.com/provenance-io/provenance/pull/384)
 * Add Marker governance proposal for setting denom metadata [#369](https://github.com/provenance-io/provenance/issues/369)
+* Add `config` command to cli for client configuration [#394](https://github.com/provenance-io/provenance/issues/394)
 
 ### Bug Fixes